:arrow_up: Upgrade deps

+++ pkg/storage/mocks/storage_driver.go
@@ -156,7 +156,7 @@ func (mr *MockStorageDriverMockRecorder) Upload(arg0, arg1, arg2 any) *gomock.Ca
 }
 
 // UploadPart mocks base method.
-func (m *MockStorageDriver) UploadPart(arg0 context.Context, arg1, arg2 string, arg3 int64, arg4 io.Reader) (string, error) {
+func (m *MockStorageDriver) UploadPart(arg0 context.Context, arg1, arg2 string, arg3 int32, arg4 io.Reader) (string, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "UploadPart", arg0, arg1, arg2, arg3, arg4)
 	ret0, _ := ret[0].(string)
